src/inference/requirements.txt
fastapi
uvicorn[standard]
pillow
transformers>=4.45
accelerate
